+++ src/independent/gateway/networkmessage.cpp
@@ -14,10 +14,10 @@
 
 namespace RODOS {
 
-size_t NetworkMessage::setUserData(const void* data, size_t len) {
+uint16_t NetworkMessage::setUserData(const void* data, uint16_t len) {
     len =  MIN(len, MAX_NETWORK_MESSAGE_LENGTH);
     memcpy(userDataC, data, len);
-    put_len(static_cast<uint16_t>(len));
+    put_len(len);
     return len;
 }
 
@@ -27,7 +27,7 @@ size_t NetworkMessage::setUserData(const void* data, size_t len) {
  * @param length of message buffer, supposed to be <= MAX_NETWORK_MESSAGE_LENGTH
  * @return length of message copied
  */
-size_t NetworkMessage::getUserData(void* destination, size_t maxLen) {
+uint16_t NetworkMessage::getUserData(void* destination, uint16_t maxLen) {
     maxLen = MIN(maxLen, get_len());
     memcpy(destination, userDataC, maxLen);
     return maxLen;
